Rename Associaiton List to Association Map, to make it more clear.
