+++ lib/AssociationMap.js
@@ -2,6 +2,7 @@ var _ = require('lodash');
 var Promise = require('bluebird');
 
 function AssociationList(associations) {
+
     var self = {};
 
     function evaluateNestedAssociations(definition, ctx, request, components) {
