+++ lib/associationProcessorService.js
@@ -3,7 +3,7 @@
 /**
  * Created by jcsokolow on 5/3/16.
  */
-var AssociationList = require("./AssociationList");
+var AssociationList = require("./AssociationMap");
 /**
  * This function processes child associations before parents.
  */
